Allow to process requests in parallel by not using temporary file with the same name.
When running the plugging using workers as in:

```bash
uvicorn server.main:app --workers 8
```

And it starts processing files concurrently, there are clashes because
the workers where using the same temporary file `/tmp/temp_file`. Also
one can't run two retrieval plugins in the same machine.

To avoid this here we ask the operating system to create a random
temporary file for us and this method is also operating system
independent.

Does not assume that the `/tmp` directory exists and is writable.